+++ test/typescript/apis/project/cds-server.ts
@@ -13,8 +13,8 @@ cds.middlewares.add(h, {before: 'foo', after: 'foo'})
 cds.middlewares.add(h, {after: 'foo', at: 2})
 
 cds.error(404, 'Not Found')
-cds.error('Not Found', {code: 404, stack: 'foo'})
+cds.error('Not Found', {status: 404, stack: 'foo'})
 cds.error `Message with formatted: ${{foo: 'bar'}}`
-cds.error({code: 409, message: 'ERROR_MESSAGE'})
+cds.error({status: 409, message: 'ERROR_MESSAGE'})
 
 cds.app satisfies Application