+++ CHANGELOG.md
@@ -7,10 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [Unreleased]
 
 ### Added
+- Types for `SELECT.stream()`
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
+- Types for `req.error()` et al.
+- Types for `cds.error()`
 ### Security
 
 ## [0.15.0] - 2025-09-26