+++ CHANGELOG.md
@@ -8,9 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Added
 - Types for `SELECT.stream()`
+- `service.actions`
 ### Changed
 ### Deprecated
+- `service.operations`
 ### Removed
+- Undocumented `service.entities()`, `service.events()`, `service.types()`, and `service.operations()`
 ### Fixed
 - Types for `req.error()` et al.
 - Types for `cds.error()`