+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Fixed
 - Types for `req.error()` et al.
 - Types for `cds.error()`
+- Type for `cds.middlewares.before`
 ### Security
 
 ## [0.15.0] - 2025-09-26