+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [Unreleased]
 
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.16.0] - 2026-02-09
+
+### Added
 - Added `Request.messages`, `Request.errors` and `Request.results`
 - Types for `SELECT.stream()`
 - `service.actions`